requirements.txt
aiohttp
pytest  
pytest-aiohttp 
pytest-cov 
allure-pytest
